+++ analyzer/lookup.c
@@ -169,3 +169,4 @@ void fillLookups(void)
     }
   }
 }
+